Added initial patches
Extend patches and core patch are now provided as separate files and
download zip button can be used if an archive is desired. The extra
patches are still packaged as an archive since they are just extras
provided for anyone that needs them, but are unsupported from STEP 2.2.7
and up.